v2_learning_system_real/requirements.txt
aiohttp>=3.8.0
openai>=1.0.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
